+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.8.0] - Not released
+### Added
+- New Datadog counters for invitations:'invitation.create-time',
+  'invitation.create-requests', and 'invitation.use-requests'.
+
 ### Changed
 - Invitations from inactive (i.e. in some 'gone' status) users stop working.